release 0.1.0

+++ database/migrations/2020_05_27_000112_create_page_user_table.php
@@ -16,7 +16,7 @@ class CreatePageUserTable extends Migration
         Schema::create('page_user', function (Blueprint $table) {
             $table->increments('id');
             $table->unsignedInteger('user_id');
-            $table->unsignedInteger('page_id');
+            $table->unsignedBigInteger('page_id');
             $table->timestamps();
 
             $table->foreign('user_id')->references('id')->on('users');
